refactor(sql-playground): Step 6 - Move dialogs and enable lazy loading
- Moved 3 dialog components to features/sql-playground/dialogs/
  - new-db-dialog, new-sql-template, share-playground-link-dialog
- Updated dialog imports in db-control components (6 → 3 levels)
- Added dialog declarations to SqlPlaygroundFeatureModule
- Converted to lazy-loaded route with loadChildren in app-routing.module.ts
- Added route definition to sql-playground-routing.module.ts (path: '', component: SqlPlaygroundComponent)
- Updated tsconfig.json module from es2015 to es2020 for dynamic imports
- Removed dialog imports/declarations from app.module.ts
- Added StoreModule.forRoot() and EffectsModule.forRoot() to app.module.ts
- Removed forRoot() calls from feature module (Store, Effects, I18NextModule, StoreDevtools)
- Added I18NextModule to feature module imports (without forRoot)
- Fixed SCSS import path in new-sql-template.component.scss (3 → 6 levels)
- Deleted old page-components/sql-playground/ directory

SQL Playground is now fully isolated and lazy-loaded.
Build: ✅ SUCCESS | Runtime: ✅ WORKING
