+++ modules/fbs-core/web/src/app/app.module.ts
@@ -62,7 +62,6 @@ import { MenuBarComponent } from "./tool-components/menu-bar/menu-bar.component"
 import { NewCheckerDialogComponent } from "./dialogs/new-checker-dialog/new-checker-dialog.component";
 import { ConfirmDialogComponent } from "./dialogs/confirm-dialog/confirm-dialog.component";
 import { TextConfirmDialogComponent } from "./dialogs/text-confirm-dialog/text-confirm-dialog.component";
-import { NewDbDialogComponent } from "./dialogs/new-db-dialog/new-db-dialog.component";
 import {
   NgxMatDatetimePickerModule,
   NgxMatNativeDateModule,
@@ -86,22 +85,20 @@ import { SqlCheckerResultsComponent } from "./page-components/sql-checker/sql-ch
 import { MatTableModule } from "@angular/material/table";
 import { MatSortModule } from "@angular/material/sort";
 import { AnalyticsToolComponent } from "./page-components/analytics-tool/analytics-tool.component";
-import { NewSqlTemplateComponent } from "./dialogs/new-sql-template/new-sql-template.component";
 import { ExportTasksDialogComponent } from "./dialogs/export-tasks-dialog/export-tasks-dialog.component";
 import "mathlive";
 import "@cortex-js/compute-engine";
 import { MathInputComponent } from "./tool-components/math-input/math-input.component";
-import { SharePlaygroundLinkDialogComponent } from "./dialogs/share-playground-link-dialog/share-playground-link-dialog.component";
 import { FbsModellingComponent } from "./page-components/fbs-modelling/fbs-modelling.component";
 import { FbsQuestionaryComponent } from "./page-components/fbs-questionary/fbs-questionary.component";
 import { I18NextModule } from "angular-i18next";
 import { I18N_PROVIDERS } from "./util/i18n";
 import { LanguageMenuComponent } from "./page-components/sidebar/language-menu/language-menu.component";
+import { StoreModule } from "@ngrx/store";
+import { EffectsModule } from "@ngrx/effects";
 import { registerLocaleData } from "@angular/common";
 import localeDe from "@angular/common/locales/de";
 import localeDeExtra from "@angular/common/locales/extra/de";
-import { SqlPlaygroundModule } from "./page-components/sql-playground/sql-playground.module";
-import { SqlPlaygroundFeatureModule } from "./features/sql-playground/sql-playground.module"; // Step 2: New feature module shell
 import { MyGroupsComponent } from "./page-components/my-groups/my-groups.component";
 import { GroupSelectionComponent } from "./page-components/course-detail/group-selection/group-selection.component";
 import { NewGroupDialogComponent } from "./dialogs/new-group-dialog/new-group-dialog.component";
@@ -192,11 +189,8 @@ export const httpInterceptorProviders = [
     SqlCheckerResultsComponent,
     TextConfirmDialogComponent,
     AnalyticsToolComponent,
-    NewDbDialogComponent,
-    NewSqlTemplateComponent,
     ExportTasksDialogComponent,
     MathInputComponent,
-    SharePlaygroundLinkDialogComponent,
     FbsModellingComponent,
     FbsQuestionaryComponent,
     LanguageMenuComponent,
@@ -245,8 +239,8 @@ export const httpInterceptorProviders = [
     MatTableModule,
     MatSortModule,
     I18NextModule.forRoot(),
-    SqlPlaygroundModule,
-    SqlPlaygroundFeatureModule, // Step 2: Import new feature module shell (alongside old module temporarily)
+    StoreModule.forRoot({}),
+    EffectsModule.forRoot([]),
     AngularEditorModule,
   ],
   entryComponents: [
