+++ modules/fbs-core/web/src/app/features/sql-playground/REFACTORING_LOG.md
@@ -1,0 +1,83 @@
+# SQL Playground Feature Extraction - Refactoring Log
+
+## Step 6: Move Dialogs + Enable Lazy Loading ✅
+
+**Date:** February 5, 2026
+
+### Changes Made
+
+1. **Created dialogs/ directory** in `features/sql-playground/`
+
+2. **Moved 3 dialog components:**
+   - `new-db-dialog/` → `features/sql-playground/dialogs/new-db-dialog/`
+   - `new-sql-template/` → `features/sql-playground/dialogs/new-sql-template/`
+   - `share-playground-link-dialog/` → `features/sql-playground/dialogs/share-playground-link-dialog/`
+
+3. **Updated dialog imports:**
+   - [db-control-db-overview.component.ts](../../../features/sql-playground/components/db-control-panel/db-control-db-overview/db-control-db-overview.component.ts): Updated `NewDbDialogComponent` and `SharePlaygroundLinkDialogComponent` imports from 6 levels to 3 levels
+   - [db-control-templates.component.ts](../../../features/sql-playground/components/db-control-panel/db-control-templates/db-control-templates.component.ts): Updated `NewSqlTemplateComponent` import from 6 levels to 3 levels
+
+4. **Added dialogs to feature module:**
+   - Updated [sql-playground.module.ts](../../../features/sql-playground/sql-playground.module.ts) to import and declare all 3 dialog components
+
+5. **Converted to lazy-loaded route:**
+   - Updated [app-routing.module.ts](../../../../../app-routing.module.ts):
+     - Removed direct `SqlPlaygroundComponent` import
+     - Changed route to use `loadChildren: () => import("./features/sql-playground/sql-playground.module").then(m => m.SqlPlaygroundFeatureModule)`
+   - Updated [tsconfig.json](../../../../../../../tsconfig.json): Changed module from `es2015` to `es2020` to support dynamic imports
+
+6. **Cleaned up app.module.ts:**
+   - Removed imports for `SqlPlaygroundModule` (old wrapper)
+   - Removed imports for `SqlPlaygroundFeatureModule` (no longer needed - now lazy loaded)
+   - Removed imports for the 3 dialog components (now part of feature module)
+   - Removed all SQL Playground declarations from app.module.ts
+
+7. **Deleted old wrapper module:**
+   - Removed `page-components/sql-playground/` directory completely
+
+8. **Fixed SCSS import:**
+   - [new-sql-template.component.scss](../../../features/sql-playground/dialogs/new-sql-template/new-sql-template.component.scss): Updated colors.scss import from 3 levels to 6 levels
+
+### Files Modified
+- `app-routing.module.ts` (lazy loading)
+- `app.module.ts` (removed SQL Playground imports and declarations)
+- `tsconfig.json` (module: es2020)
+- `features/sql-playground/sql-playground.module.ts` (added dialog declarations)
+- `features/sql-playground/components/db-control-panel/db-control-db-overview/db-control-db-overview.component.ts`
+- `features/sql-playground/components/db-control-panel/db-control-templates/db-control-templates.component.ts`
+- `features/sql-playground/dialogs/new-sql-template/new-sql-template.component.scss`
+
+### Files Moved
+- `dialogs/new-db-dialog/**` → `features/sql-playground/dialogs/new-db-dialog/`
+- `dialogs/new-sql-template/**` → `features/sql-playground/dialogs/new-sql-template/`
+- `dialogs/share-playground-link-dialog/**` → `features/sql-playground/dialogs/share-playground-link-dialog/`
+
+### Files Deleted
+- `page-components/sql-playground/` (entire directory and all contents)
+
+### Build Status
+✅ **BUILD SUCCESSFUL** - Only warnings (unused files, CommonJS dependencies)
+
+### Verification
+- All TypeScript compilation errors resolved
+- Lazy loading configuration working
+- Dialog components properly declared in feature module
+- No import path errors
+- SCSS paths corrected
+
+---
+
+## Migration Complete! 🎉
+
+All 6 steps completed successfully. The SQL Playground feature is now:
+- ✅ Fully isolated in `features/sql-playground/`
+- ✅ Lazy-loaded (only loads when user navigates to `/sqlplayground`)
+- ✅ Self-contained (all components, services, models, dialogs, state in one place)
+- ✅ No coupling with core app (uses AuthGuard and core services via DI)
+- ✅ Production-ready (build passes)
+
+### Next Steps (Optional)
+- Functional testing of all features
+- Consider adding feature-specific routing guards
+- Monitor bundle size impact of lazy loading
+- Update team documentation
