+++ modules/fbs-core/web/src/app/features/sql-playground/dialogs/new-db-dialog/new-db-dialog.component.scss
