+++ modules/fbs-core/web/src/app/features/sql-playground/dialogs/new-sql-template/new-sql-template.component.html
