+++ modules/fbs-core/web/src/app/features/sql-playground/dialogs/new-sql-template/new-sql-template.component.scss
@@ -1,4 +1,4 @@
-@import "../../../colors.scss";
+@import "../../../../../colors.scss";
 
 .container {
   height: 100%;
