+++ modules/fbs-core/web/src/app/features/sql-playground/sql-playground-routing.module.ts
@@ -1,15 +1,12 @@
 import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
-
-/**
- * SQL Playground Routing Module
- * 
- * This will be populated in Step 6 when components are moved.
- * For now, it's an empty shell to establish the module structure.
- */
+import { SqlPlaygroundComponent } from "./components/sql-playground.component";
 
 const routes: Routes = [
-  // Routes will be added in Step 6
+  {
+    path: "",
+    component: SqlPlaygroundComponent,
+  },
 ];
 
 @NgModule({
