+++ modules/fbs-core/web/tsconfig.json
@@ -5,7 +5,7 @@
     "outDir": "./dist/out-tsc",
     "sourceMap": true,
     "declaration": false,
-    "module": "es2015",
+    "module": "es2020",
     "moduleResolution": "node",
     "experimentalDecorators": true,
     "target": "es2020",
